Separate `ir` module from `codegen` into `sonatina-ir` crate
