+++ crates/codegen/src/lib.rs
@@ -5,10 +5,6 @@
 pub mod cfg;
 pub mod critical_edge;
 pub mod domtree;
-pub mod ir;
-pub mod isa;
 pub mod loop_analysis;
 pub mod optim;
 pub mod post_domtree;
-
-pub use isa::TargetIsa;
