+++ crates/ir/src/builder/mod.rs
@@ -7,37 +7,31 @@ pub use module_builder::ModuleBuilder;
 
 pub use ssa::Variable;
 
-#[cfg(test)]
-pub(crate) mod test_util {
+pub mod test_util {
     use super::*;
 
     use sonatina_triple::TargetTriple;
 
     use crate::{
-        ir::{
-            ir_writer::FuncWriter,
-            module::{FuncRef, Module},
-            Function, Linkage, Signature, Type,
-        },
+        ir_writer::FuncWriter,
         isa::{IsaBuilder, TargetIsa},
+        module::{FuncRef, Module},
+        Function, Linkage, Signature, Type,
     };
 
-    pub(crate) fn build_test_isa() -> TargetIsa {
+    pub fn build_test_isa() -> TargetIsa {
         let triple = TargetTriple::parse("evm-ethereum-london").unwrap();
         IsaBuilder::new(triple).build()
     }
 
-    pub(crate) struct TestModuleBuilder {
+    pub struct TestModuleBuilder {
         module_builder: ModuleBuilder,
         func_ref: Option<FuncRef>,
     }
 
     impl TestModuleBuilder {
-        pub(crate) fn new() -> Self {
-            Self {
-                module_builder: ModuleBuilder::new(build_test_isa()),
-                func_ref: None,
-            }
+        pub fn new() -> Self {
+            Self::default()
         }
 
         pub fn func_builder(&mut self, args: &[Type], ret_ty: &Type) -> FunctionBuilder {
@@ -52,8 +46,17 @@ pub(crate) mod test_util {
         }
     }
 
-    pub(crate) fn dump_func(func: &Function) -> String {
+    pub fn dump_func(func: &Function) -> String {
         let mut writer = FuncWriter::new(func);
         writer.dump_string().unwrap()
+    }
+
+    impl Default for TestModuleBuilder {
+        fn default() -> Self {
+            Self {
+                module_builder: ModuleBuilder::new(build_test_isa()),
+                func_ref: None,
+            }
+        }
     }
 }
