+++ crates/parser/Cargo.toml
@@ -6,7 +6,7 @@ edition = "2021"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-sonatina-codegen = { path = "../codegen" }
+sonatina-ir = { path = "../ir" }
 sonatina-triple = { path = "../triple" }
 smallvec = "1.7.0"
 cranelift-entity = "0.77"
